+++ src/presolve/HPresolve.cpp
@@ -3175,7 +3175,7 @@ HPresolve::Result HPresolve::rowPresolve(HighsPostsolveStack& postsolve_stack,
             if (x1Cand != -1) {
               HighsInt x1Pos = rowpositions[x1Cand];
               HighsInt x1 = Acol[x1Pos];
-              double rhs2 = rhs / d;
+              double rhs2 = rhs / double(d);
               if (std::abs(std::round(rhs2) - rhs2) <=
                   mipsolver->mipdata_->epsilon) {
                 // the right hand side is integral, so we can substitute
@@ -3188,7 +3188,7 @@ HPresolve::Result HPresolve::rowPresolve(HighsPostsolveStack& postsolve_stack,
               } else {
                 // we can substitute x1 = d * z + b, with b = a1^-1 rhs (mod d)
 
-                // first compute the modular multiplciative inverse of a1^-1
+                // first compute the modular multiplicative inverse of a1^-1
                 // (mod d) of a1
                 int64_t a1 = std::round(intScale * Avalue[x1Pos]);
                 a1 = HighsIntegers::mod(a1, d);
@@ -3206,8 +3206,8 @@ HPresolve::Result HPresolve::rowPresolve(HighsPostsolveStack& postsolve_stack,
                 // we directly fix x1 instead of first substituting with d * z +
                 // b.
                 double zLower =
-                    std::ceil((model->col_lower_[x1] - b) / d - primal_feastol);
-                double zUpper = std::floor((model->col_upper_[x1] - b) / d +
+                    std::ceil((model->col_lower_[x1] - b) / double(d) - primal_feastol);
+                double zUpper = std::floor((model->col_upper_[x1] - b) / double(d) +
                                            primal_feastol);
 
                 if (zLower == zUpper) {
